flytter til core
