+++ @navikt/core/react/src/index.ts
@@ -15,6 +15,7 @@ export * from "./panel";
 export * from "./popover";
 export * from "./speech-bubble";
 export * from "./tag";
+export * from "./table";
 export * from "./typography";
 export * from "./util";
