+++ @navikt/core/react/src/table/index.ts
