+++ @navikt/internal/css/index.css
@@ -1,4 +1,3 @@
 @import "divider.css";
 @import "dropdown.css";
 @import "header.css";
-@import "table.css";
