+++ @navikt/internal/react/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./divider";
 export * from "./dropdown";
 export * from "./header";
-export * from "./table";
